Removed couch_users.py
