+++ corehq/apps/users/dbaccessors/__init__.py
@@ -2,8 +2,6 @@ from .all_commcare_users import (
     get_all_commcare_users_by_domain,
     get_practice_mode_mobile_workers,
     get_user_docs_by_username,
-)
-from .couch_users import (
     get_user_id_and_doc_type_by_domain,
     get_user_id_by_username,
 )
